+++ src/main/java/com/matchday/matchdayserver/team/model/entity/Team.java
@@ -1,6 +1,6 @@
 package com.matchday.matchdayserver.team.model.entity;
 
-import com.matchday.matchdayserver.teamuser.model.entity.UserTeam;
+import com.matchday.matchdayserver.userteam.model.entity.UserTeam;
 import jakarta.persistence.*;
 import lombok.Builder;
 import lombok.Getter;
